Make GitHubStubClient's PR state consistent at the end of a merge
Within tests, GitHubStubClient still has an out of date view of the PR
state after a merge because nothing explicitly asks for the PRs and
autoClosePrs doesn't get a chance to get invoked. I could have it
triggered by all calls to create or update PRs as well, but it's already
very slow when running with the CliGitClient, so I'd rather have an
explicit call to it within the merge function and promote it to the
interface and have the production implementation do nothing.

commit-id: I2a224564