Test fix: avoid undefined behavior in autoClosePrs
I should not be using logAll() here, especially not:
```
localGit.logAll().associateBy(Commit::id)
```
Since there could be multiple commits with the same ID in the tree and
the ordering of logAll() is not defined.

Instead I'm using the remote branches, which are unambiguous.

(This fixes a spurious test failure in
`GitJasprTest.merge rebases remaining PRs after merge`)

commit-id: I768b7168