`clean` prunes stale tracking branches
In the scenario where someone merges from another working copy, it's
possible for `git branch -r` to return a list containing branches that
no longer exist in the remote. This is mostly harmless, but it causes
`git jaspr clean` to include them in its output as orphaned branches,
which can be confusing. This change prunes the given remote before
getting the list of branches to avoid this scenario.

commit-id: I1ecf919d