Log file path is logged to console when using DEBUG or lower
The initialization order here was a bit turned around. We were logging
the log file path (at DEBUG level), but doing so before fully
configuring logging from the CLI options. Since the bootstrap logging
level was INFO for the console appender, this meant the log file path
was being logged only to the actual log file itself... which isn't so
useful if you don't know where it is to begin with. :)

Now I apply all the log file options first, then log the path. This
means if you invoke the command with `-lDEBUG` (f.e.), you'll see the
path.

This also has the side effect of fixing a hack I was using for the
external process test... the log line containing the log file path was
being picked up as part of the output since the log configuration for
the test source set defaults to TRACE level, and this log configuration
is what's getting baked in when running the external process. Updating
the init order here means the default threshold filter of INFO level is
active, which suppresses the log file path line from the output of the
external process. Happy accident.

commit-id: Id21b9603