Improve output when branch cleanup requires retrying
If we get a transient error during branch cleanup and have to retry,
it's easy to assume the whole operation failed since the last thing you
see if the stack trace from the previous failed attempt. To mitigate
this somewhat, we print a success message after branch cleanup only if
it required 1 or more retries.

commit-id: Ic2fe11af