Hotfix for "A pull request already exists" error
The getPullRequests.graphql query contains a hardcoded limit on how many
PRs to fetch. It was set to 60 for some reason, and when using with a
repo that contains more than that number, it's possible for a PR in your
stack to not be in the response and therefore make Jaspr think it needs
to push a new one.

The proper fix for this is to paginate the PRs, but for now I'm bumping
up the limit to 120. I will revisit and fix properly soon.

commit-id: Ia115de4b