GitHubTestHarness: properly handle creating "sibling" branches
The test harness was not properly creating a repo structure with
"sibling" branches like B, C, and D in this example:

* a9b9cdf - 2
| * 5d41f9d - (C) 1.2
|/
| * c5a2e1c - (D) 1.3
|/
| * 7b83731 - (B) 1.1
|/
* e162c52 - 1
* 8543836 - (main) Initial commit

It was creating B, C, and D as children in the same branch. After
recursing into a branch, createCommits() needed to do a checkout
to reset the repo state before creating the next one.

commit-id: I57638d30