Fix getRemoteBranches behavior with newer versions of git
At some point, the git CLI did not return "$remoteName/HEAD" when
executing "git branch -r" and now it does for some reason. This breaks
some tests, so let's filter it out in both clients and add tests to
ensure the filtering works.

Also, fix a couple of getRemoteBranches tests that were passing by
coincidence due to not specifying the correct remote name and therefore
getting empty lists.

Also, fix an issue where JGit is simplying author email to just
the TLD. I don't know what that's about but it's not a big dea.

Also, update pull_requests.yml to run-on `ubuntu-latest`

commit-id: I1011dd0b