Avoid class initialization error during graal native image build
The class kotlin.DeprecationLevel ends up getting initialized at build
time. During compilation, graalvm presumably attempts to see if a class
is annotated with com.oracle.svm.core.hub.Hybrid, which has the side
effect of initializing kotlin.DeprecationLevel. This causes the build to
fail if we don't explicitly request that it be initialized at build
time, which I do in this commit.

commit-id: Id09bee48