Update merge behavior to prevent GH from showing merged PRs as closed
When merging a stack, jaspr "rolls up" all commits to be merged into
the last mergeable PR and merges just that PR. Previously, Jaspr would
close the PRs being rolled up and delete their branches. However, this
would result in GitHub showing those PRs as closed and not merged,
which is misleading.

Now Jaspr no longer closes the PRs. It still deletes the branches,
but it delays for a short time (2 seconds) to allow GitHub to "notice"
that the commits on these PR branches are in their target branches
and to mark those PRs as merged as well. In my limited testing, 2
seconds appeared to be enough time, but this delay may need to be
increased if it proves unreliable.

commit-id: I48fcc46c